Fixes corner case in CssAbsolute:  (Using a Cached S3 storage AND the file to concatenate isn't available on the local filesystem) which happens when the results of collectstatic aren't available locally for the request.
The fallback system was working, files were being included individually, but concatenation wasn't happening reliably.

The corner case:
E.G. This happens on Heroku when `DISABLE_COLLECTSTATIC` is set.  When DISABLE_COLLECTSTATIC is set, the local cache results of collectstatic are put on a worker node which is destroyed at the end of the command and the collected files aren't added to the compressed application slug which will be used to spin up new dynos.

This manifestation on heroku is a specific case, the general case has the ability to arise anywhere the local filesystem is treated as ephemeral (especially if there are >1 servers).